+++ src/LIN3S/AdminBundle/Controller/AdminController.php
@@ -11,8 +11,7 @@
 
 namespace LIN3S\AdminBundle\Controller;
 
-use LIN3S\AdminBundle\Annotation\EntityConfiguration as EntityConfigurationAnnotation;
-use LIN3S\AdminBundle\Configuration\EntityConfiguration;
+use LIN3S\AdminBundle\Configuration\Model\Entity;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -24,22 +23,21 @@ use Symfony\Component\HttpFoundation\Request;
 class AdminController extends Controller
 {
     /**
-     * @EntityConfigurationAnnotation()
-     *
      * List action.
      *
-     * @param string              $entity       The entity name
-     * @param EntityConfiguration $entityConfig The entity configuration
-     * @param Request             $request      The request
+     * @param string  $entity       The entity name
+     * @param Entity  $entityConfig The entity configuration
+     * @param Request $request      The request
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function listAction($entity, EntityConfiguration $entityConfig, Request $request)
+    public function listAction($entity, Request $request)
     {
+        $entityConfig = $this->get('lin3s_admin.configuration.factory.entity')->createFor($entity);
         $entities = $this->get('lin3s_admin.repository')->findByRequest($request, $entityConfig);
         $totalCount = $this->get('lin3s_admin.repository')->countAll($request, $entityConfig);
 
-        return $this->render('@LIN3SAdmin/Admin/list.html.twig', [
+        return $this->render('@Lin3sAdmin/Admin/list.html.twig', [
             'entities'     => $entities,
             'entityConfig' => $entityConfig,
             'totalCount'   => $totalCount,
@@ -47,20 +45,18 @@ class AdminController extends Controller
     }
 
     /**
-     * @EntityConfigurationAnnotation()
-     *
      * Custom action.
      *
-     * @param string              $entity       The entity name
-     * @param string              $action       The action name
-     * @param string              $id           The id of the object to be edited.
-     * @param EntityConfiguration $entityConfig The entity configuration
-     * @param Request             $request      The request
+     * @param string  $entity       The entity name
+     * @param string  $action       The action name
+     * @param string  $id           The id of the object to be edited.
+     * @param Request $request      The request
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function customAction($entity, $action, $id = null, EntityConfiguration $entityConfig, Request $request)
+    public function customAction($entity, $action, $id = null, Request $request)
     {
+        $entityConfig = $this->get('lin3s_admin.configuration.factory.entity')->createFor($entity);
         $entityObject = null;
         if ($id) {
             $manager = $this->getDoctrine()->getRepository($entityConfig->className());