+++ CHANGELOG.md
@@ -5,6 +5,28 @@ This changelog references the relevant changes done between versions.
 To get the diff for a specific change, go to https://github.com/LIN3S/AdminBundle/commit/XXX where XXX is the change hash 
 To get the diff between two versions, go to https://github.com/LIN3S/AdminBundle/compare/v0.2.0...v0.3.0
 
+* 0.5.0
+    * [BC break] Refactored the whole js and scss files.
+    * [BC break] Changed LIN3SAdminBundle to Lin3sAdminBundle so, be caution with `lin3_s`
+    configuration, now it is `lin3s_admin`
+    * Added fr translation messages
+    * Refactored and unified registries
+    * [BC break] Made configuration, dependency injection, extension, registry final.
+    * [BC break] `lin3s_admin.action`, `lin3s_admin.list_field` and `lin3s_admin.list_filter` tags now require an `alias`
+    * [BC break] In config, `class` was replaced by `type` and now config type `alias` must be used instead service name
+     for action, list field and list filters. Check [reference_configuration](docs/reference_configuration.md) for further info.
+    * [BC break] Moved configuration type implementations to `Extension` folder
+    * [BC break] Full change of configuration namespaces
+    
+| Previous                                            | New                                                   |
+| --------                                            | ---                                                   |
+| LIN3S\AdminBundle\Action\Action                     | LIN3S\AdminBundle\Configuration\Model\Action          |
+| LIN3S\AdminBundle\Action\ActionType                 | LIN3S\AdminBundle\Configuration\Type\ActionType       |
+| LIN3S\AdminBundle\Action\ActionType                 | LIN3S\AdminBundle\Configuration\Type\ActionType       |
+| LIN3S\AdminBundle\Configuration\EntityConfiguration | LIN3S\AdminBundle\Configuration\Model\Entity          |
+| LIN3S\AdminBundle\ListField\ListFieldType           | LIN3S\AdminBundle\Configuration\Type\ListFieldType    |
+| LIN3S\AdminBundle\ListFilter\ListFilter             | LIN3S\AdminBundle\Configuration\Model\ListFilter      |
+| LIN3S\AdminBundle\ListFilter\ListFilterType         | LIN3S\AdminBundle\Configuration\Type\ListFilterType   |    
 * 0.4.1
     * Removed Sylius registry dependency
 * 0.4.0